flask/app/requirements.txt
uwsgi
flask
pymongo
dnspython
requests
ipaddress
apscheduler
pytz
